# docs/swagger.yaml
definitions:
...
  models.DashboardStats:
    properties:
      active_projects:
        type: integer
      attendance_status:
        type: string
      completion_rate:
        type: integer
      pending_tasks:
        type: integer
    type: object
  models.HomeDashboardResponse:
    properties:
      greeting:
        type: string
      stats:
        $ref: '#/definitions/models.DashboardStats'
      team_members:
        items:
          $ref: '#/definitions/models.TeamMember'
        type: array
      user_avatar:
        type: string
      user_name:
        type: string
    type: object
  models.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
...
  models.TeamMember:
    properties:
      avatar_url:
        type: string
      id:
        type: integer
      name:
        type: string
      position:
        type: string
    type: object
...
  models.UserRegistrationRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
      phone:
        type: string
      role:
        type: string
      username:
        type: string
    required:
    - email
    - name
    - password
    - username
    type: object
...
info:
  contact: {}
paths:
...
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and return JWT token
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      summary: User login
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Create a new user account with complete profile data
      parameters:
      - description: User registration data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UserRegistrationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Register a new user
      tags:
      - auth
  /dashboard/stats:
    get:
      description: Get overview statistics for dashboard
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.DashboardStats'
      security:
      - BearerAuth: []
      summary: Get dashboard statistics
      tags:
      - dashboard
  /home:
    get:
      description: Get complete data for home dashboard including stats and team members
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.HomeDashboardResponse'
      security:
      - BearerAuth: []
      summary: Get home dashboard data
      tags:
      - dashboard
...
  /teams:
    get:
      description: Get list of team members
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.TeamMember'
            type: array
      security:
      - BearerAuth: []
      summary: Get team members
      tags:
      - dashboard
...
swagger: "2.0"